refactor: consolidate gateway module into single file
Move all gateway types, handlers, and helpers into mod.rs to match
the pattern used by file_explorer.rs and file_viewer.rs. Reuse the
existing format_bytes function from file_viewer instead of duplicating.
